Se corrigen paths
